+++ src/vscode.proposed.editorInsets.d.ts
@@ -1,0 +1,24 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+declare module 'vscode' {
+  export interface WebviewEditorInset {
+    readonly editor: TextEditor;
+    readonly line: number;
+    readonly height: number;
+    readonly webview: Webview;
+    readonly onDidDispose: Event<void>;
+    dispose(): void;
+  }
+
+  export namespace window {
+    export function createWebviewTextEditorInset(
+      editor: TextEditor,
+      line: number,
+      height: number,
+      options?: WebviewOptions,
+    ): WebviewEditorInset;
+  }
+}